abi: catch the mirror up to v35, and mark the dead file-based mutators -- #308
The engine side of A1-C3 and A1-C9 landed this session; this is the half that makes
them real. Doing it in the same breath IS A1-C4 -- the mirror sitting a version
behind engine HEAD was itself a confirmed finding, so deferring this would recreate
the bug the other commit fixes.

- AbiVersion 34 -> 35.
- WzEditorAssetGraphSnapshotAbi.Reserved becomes AbiVersion, at the same offset and
  size, so every existing layout assert still holds; a new AssertOffset pins it.
- ReadAssetGraphSnapshot now calls ValidateAbiVersion. It was the ONE top-level
  reader that never did, which is what let a future graph-struct bump skew the graph
  editor's stride silently while every other reader failed loudly.
- SetNodeCollision's asset-graph node id widens uint -> ulong through the whole
  chain: P/Invoke, client, interface, session. It was the only verb in its family
  carrying a 32-bit id into a 64-bit field. The interface change surfaced the test
  fake immediately (CS0535 on RecordingEditorSession), which is the signature working
  as intended -- its method and recording tuple widen with it.

A1-C12 also lands here, since it edits the same interface file. It was reported as
SetSceneNodeTransform forwarding nine raw strings where its live sibling parses them,
persisting a mid-edit value the preview had skipped. That scenario is unreachable:
wz_host_scene_set_node_transform casts all nine strings to (void) and returns "scene
authoring mutations are not implemented in the engine ABI yet", nothing is persisted,
and no ViewModel calls it. Adding parsing would polish a path that cannot write.

The real defect is the trap: three members (SetSceneNodeProperties,
SetSceneNodeTransform, SetSceneNodeCamera) sat unmarked beside their working *Live
twins, so choosing the wrong one by name is a silent no-op. The live ones are
documented; these were not. Kept rather than deleted -- the X / XLive pairing is the
tell that separates the two families, and removing half makes the rest look
arbitrary.

338/338 tests.

Found by the #320 adversarial audit rotation (part A1).

+++ editor/src/Wozzits.Editor.HostClient/IWozzitsEngineEditorSession.cs
@@ -80,6 +80,18 @@ public interface IWozzitsEngineEditorSession
 
     EngineMutationResponse SetAssetGraphZoom(double zoom);
 
+    // DEAD file-based twins -- do not route an edit through these. Their engine
+    // exports (wz_host_scene_set_node_properties / _node_transform / _camera) cast
+    // every value argument to (void) and return
+    // "scene authoring mutations are not implemented in the engine ABI yet", so the
+    // call always FAILS and nothing is ever written. Every working edit goes through
+    // the runtime seam instead: the *Live members below, which post to the running
+    // engine and are persisted by SaveScene.
+    //
+    // They are kept because the naming pair (SetSceneNodeX vs SetSceneNodeXLive) is
+    // the tell that distinguishes the two families, and deleting half of it would
+    // make the surviving half look arbitrary. Marked rather than deleted so the next
+    // reader does not have to trace the export to find out which one persists.
     EngineMutationResponse SetSceneNodeProperties(
         string nodeId,
         string name,
@@ -198,6 +210,9 @@ public interface IWozzitsEngineEditorSession
     // switch back). Empty response when no viewport is running.
     EngineSceneSnapshotResponse LoadRuntimeSceneSnapshot();
 
+    // DEAD file-based twin, like SetSceneNodeProperties/SetSceneNodeTransform above:
+    // wz_host_scene_set_camera discards its arguments and fails. Camera edits go
+    // through SetRuntimeSceneNodeCamera.
     EngineMutationResponse SetSceneNodeCamera(
         string nodeId,
         EngineSceneCameraEdit edit);
@@ -309,7 +324,7 @@ public interface IWozzitsEngineEditorSession
     // Live + host-gated, no-op success when no viewport is running.
     EngineMutationResponse SetNodeCollision(
         string nodeId,
-        uint assetGraphNodeId,
+        ulong assetGraphNodeId,
         bool constrainMovement);
 
     // Author a node's Atmosphere component: which Atmosphere asset-graph node the
